package the game logic seperately
- create packages for game logic and andriod dependancies
- refactor code to ensure game logic is independent of android logic
- modularise app to further reduce coupling
