--- app/src/main/java/com/example/tictactoe/Turns.kt
@@ -1,11 +1,0 @@
-package com.example.tictactoe
-
-class Turns {
-    companion object {
-        var myTurn: Boolean = false
-
-        fun switchTurns() {
-            myTurn = !myTurn
-        }
-    }
-}
