+++ settings.gradle
@@ -1,2 +1,2 @@
-include ':app'
+include ':app', ':core'
 rootProject.name='Tic Tac Toe'
